requirements.txt
# AI Bug Tracker - Python Dependencies
# Phase 2: Link verified (AI disabled for MVP)

# Web Framework
Flask==3.1.0
flask-cors==5.0.1

# Environment
python-dotenv==1.1.0

# Export
openpyxl==3.1.5

# Utilities
werkzeug==3.1.3
